requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
pymongo>=4.6.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
